Enforce use of external TLS certificate
If an external certificate is specified on the command line, validate and enforce the use of that certificate.

This change loads the certificate and key files into a Certificate object (after validation), which is then passed into the PeerManager, instead of sending the paths to these files individually. This allows checking that the files are readable and valid before syncing so that certificate errors are presented immediately and not buried in stdout noise.